openfaas-ocr/requirements.txt
Pillow
pytesseract
ocrmypdf
textract